# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "S2OEvaluationRequest": {
        "properties": {
          "access": {
            "title": "Access",
            "type": "string"
          },
          "data_classification": {
            "title": "Data Classification",
            "type": "string"
          },
          "hosting_environment": {
            "title": "Hosting Environment",
            "type": "string"
          },
          "release_type": {
            "title": "Release Type",
            "type": "string"
          },
          "solution_type": {
            "title": "Solution Type",
            "type": "string"
          }
        },
        "required": [
          "data_classification",
          "access",
          "solution_type",
          "hosting_environment",
          "release_type"
        ],
        "title": "S2OEvaluationRequest",
        "type": "object"
      },
...
      "SubmitAnswersRequest": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "title": "Answers",
            "type": "object"
          }
        },
        "title": "SubmitAnswersRequest",
        "type": "object"
      },
...
      "TokenRequest": {
        "properties": {
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "TokenRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Automated Security Assessment with LLMs & RAG",
    "title": "DocSentinel API",
    "version": "4.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Login Api V1 Auth Login Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "governance-auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "post": {
        "operationId": "me_api_v1_auth_me_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Me Api V1 Auth Me Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "governance-auth"
        ]
      }
    },
...
    "/api/v1/oscal/catalog": {
      "get": {
        "description": "Export resolved policy-pack controls in an OSCAL catalog-like shape.",
        "operationId": "export_oscal_catalog_api_v1_oscal_catalog_get",
        "parameters": [
          {
            "in": "query",
            "name": "framework_ids",
            "required": false,
            "schema": {
              "default": [],
              "items": {
                "type": "string"
              },
              "title": "Framework Ids",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Oscal Catalog Api V1 Oscal Catalog Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Oscal Catalog",
        "tags": [
          "governance-exports"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/oscal/assessment-results": {
      "get": {
        "description": "Export project controls, evidence, and findings as OSCAL assessment results.",
        "operationId": "export_project_oscal_assessment_results_api_v1_projects__project_id__oscal_assessment_results_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Project Oscal Assessment Results Api V1 Projects  Project Id  Oscal Assessment Results Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Project Oscal Assessment Results",
        "tags": [
          "governance-exports"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/questionnaire": {
      "get": {
        "operationId": "get_questionnaire_api_v1_projects__project_id__questionnaire_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Questionnaire Api V1 Projects  Project Id  Questionnaire Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Questionnaire",
        "tags": [
          "governance-questionnaire"
        ]
      }
    },
    "/api/v1/projects/{project_id}/questionnaire/answers": {
      "post": {
        "operationId": "submit_answers_api_v1_projects__project_id__questionnaire_answers_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitAnswersRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Submit Answers Api V1 Projects  Project Id  Questionnaire Answers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Answers",
        "tags": [
          "governance-questionnaire"
        ]
      }
    },
    "/api/v1/projects/{project_id}/questionnaire/progress": {
      "get": {
        "operationId": "get_questionnaire_progress_api_v1_projects__project_id__questionnaire_progress_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Questionnaire Progress Api V1 Projects  Project Id  Questionnaire Progress Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Questionnaire Progress",
        "tags": [
          "governance-questionnaire"
        ]
      }
    },
...
    "/api/v1/risk-assessment/evaluate": {
      "post": {
        "operationId": "evaluate_risk_assessment_api_v1_risk_assessment_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/S2OEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evaluate Risk Assessment Api V1 Risk Assessment Evaluate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evaluate Risk Assessment",
        "tags": [
          "governance-risk-assessment"
        ]
      }
    },
    "/api/v1/risk-assessment/reload": {
      "post": {
        "operationId": "reload_risk_ontology_api_v1_risk_assessment_reload_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reload Risk Ontology Api V1 Risk Assessment Reload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reload Risk Ontology",
        "tags": [
          "governance-risk-assessment"
        ]
      }
    },
    "/api/v1/s2o/evaluate": {
      "post": {
        "operationId": "evaluate_legacy_s2o_api_v1_s2o_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/S2OEvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Evaluate Legacy S2O Api V1 S2O Evaluate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Evaluate Legacy S2O",
        "tags": [
          "governance-risk-assessment"
        ]
      }
    },
...
  }
}